simulation of a propagating sinusoidal striking a dielectric medium

+++ fd1d/clang/test_1_4.c
@@ -11,8 +11,8 @@
 
 
 float sinusoidal(int t, float ds, float freq) {
-    float dt = ds/6e8;  /* time step (s) */
-    return sin(2*M_PI*freq*dt*t);
+    float dt = ds/6e8f;  /* time step (s) */
+    return sinf(2*M_PI*freq*dt*t);
 }
 
 
@@ -21,7 +21,7 @@ void exfield(int t, int nx, float *cb, float *ex, float *hy) {
     for (int i = 1; i < nx; i++)
         ex[i] += cb[i] * (hy[i-1] - hy[i]);
     /* put a sinusoidal wave at the low end */
-    ex[1] += sinusoidal(t, 0.01f, 700e6);
+    ex[1] += sinusoidal(t, 0.01f, 700e6f);
 }
 
 
@@ -31,14 +31,14 @@ void hyfield(int nx, float *ex, float *hy, float *bc) {
     ex[nx-1] = bc[3], bc[3] = bc[2], bc[2] = ex[nx-2];
     /* calculate the Hy field */
     for (int i = 0; i < nx-1; i++)
-        hy[i] += 0.5 * (ex[i] - ex[i+1]);
+        hy[i] += 0.5f * (ex[i] - ex[i+1]);
 }
 
 
 float *dielectric(int nx, float epsr) {
     float *cb = (float*) calloc(nx, sizeof(*cb));
     for (int i = 0; i < nx; cb[i] = 0.5f, i++);
-    for (int i = nx/2; i < nx; cb[i] = 0.5/epsr, i++);
+    for (int i = nx/2; i < nx; cb[i] = 0.5f/epsr, i++);
     return cb;
 }
 
@@ -53,9 +53,9 @@ int main() {
 
     float bc[4] = {0.0f};
 
-    float ds = 0.01;  /* spatial step (m) */
-    float dt = ds/6e8;  /* time step (s) */
-    float epsr = 4.0;  /* relative permittivity */
+    float ds = 0.01f;  /* spatial step (m) */
+    float dt = ds/6e8f;  /* time step (s) */
+    float epsr = 4.0f;  /* relative permittivity */
     float *cb = dielectric(nx, epsr);
 
     clock_t stime = clock();
